replace use of deprecated imp for importlib.utils
The imp library was deprecated in Python 3.4 and has been finally
removed of the standard library in version 3.12.
Instead, importlib needs to be used -- otherwise people with the new
version of Python won't be able to run the code.

I have also removed the imp library from ModuleRegulator as it was not
being used anyway.